change layout and data type

+++ app.py
@@ -16,14 +16,14 @@ def process():
     X = df.drop(['status', 'name'], axis=1)
     clf = GaussianNB()
     clf.fit(X, y)
-    testSample=np.array([int(request.form["MDVP:Fo"]),int(request.form["MDVP:Fhi"]),int(request.form["MDVP:Flo"])
-                         ,int(request.form["MDVP:Jitter"]),int(request.form["MDVP:Jitter(Abs)"]),int(request.form["MDVP:RAP"])
-                         ,int(request.form["MDVP:PPQ"]),int(request.form["Jitter:DDP"]),int(request.form["MDVP:Shimmer"])
-                         ,int(request.form["MDVP:Shimmer(dB)"]),int(request.form["Shimmer:APQ3"]),int(request.form["Shimmer:APQ5"])
-                         ,int(request.form["MDVP:APQ"]),int(request.form["Shimmer:DDA"]),int(request.form["NHR"]),
-                         int(request.form["HNR"]),int(request.form["RPDE"]),int(request.form["DFA"])
-                         ,int(request.form["spread1"]),int(request.form["spread2"]),int(request.form["D2"]),
-                         int(request.form["PPE"])])
+    testSample=np.array([float(request.form["MDVP:Fo"]),float(request.form["MDVP:Fhi"]),float(request.form["MDVP:Flo"])
+                         ,float(request.form["MDVP:Jitter"]),float(request.form["MDVP:Jitter(Abs)"]),float(request.form["MDVP:RAP"])
+                         ,float(request.form["MDVP:PPQ"]),float(request.form["Jitter:DDP"]),float(request.form["MDVP:Shimmer"])
+                         ,float(request.form["MDVP:Shimmer(dB)"]),float(request.form["Shimmer:APQ3"]),float(request.form["Shimmer:APQ5"])
+                         ,float(request.form["MDVP:APQ"]),float(request.form["Shimmer:DDA"]),float(request.form["NHR"]),
+                         float(request.form["HNR"]),float(request.form["RPDE"]),float(request.form["DFA"])
+                         ,float(request.form["spread1"]),float(request.form["spread2"]),float(request.form["D2"]),
+                         float(request.form["PPE"])])
     
     dfTest = pd.DataFrame(testSample.reshape(1,-1), columns=['MDVP:Fo(Hz)', 'MDVP:Fhi(Hz)', 'MDVP:Flo(Hz)', 'MDVP:Jitter(%)', 
                                                          'MDVP:Jitter(Abs)', 'MDVP:RAP', 'MDVP:PPQ', 'Jitter:DDP',
